Update README.md
It is shameless, but I thought this would help noobs like me. It is quasi copy and paste your remark on twitter:
'When you run "brisk meh" it creates a new directory called "meh", and puts a project in there. If you go in there and use "swift run" it will run from the command line; if you use "swift build" you'll get a finished binary you can put anywhere'.

This did help me a lot so I thought why not put it in there too.. 😅